requirements.txt
httpx[http2]
tenacity
pydantic
supabase
python-dotenv
asyncio
pytest